+++ api/src/main/java/au/org/aodn/nrmn/restapi/dto/site/SiteListItem.java
@@ -16,7 +16,7 @@ public class SiteListItem {
     String locationName;
     String state;
     String country;
-    Float latitude;
-    Float longitude;
+    Double latitude;
+    Double longitude;
     Boolean isActive;
 }
